add validation and sql scripts
