+++ src/main/java/com/olgasemenova/mirowidgets/MiroWidgetsApplication.java
@@ -3,7 +3,7 @@ package com.olgasemenova.mirowidgets;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = {"controller","service","model", "repository"})
 public class MiroWidgetsApplication {
 
 	public static void main(String[] args) {
